Another fix for conversion

+++ sdk/core/core/src/az_json_parser.c
@@ -174,7 +174,7 @@ AZ_NODISCARD static az_result az_span_reader_get_json_number_int(
   // read an integer part of the number
   while (true)
   {
-    uint64_t d = c - '0';
+    uint64_t d = (uint64_t)(c - '0');
     if (p_n->value <= (AZ_DEC_NUMBER_MAX - d) / 10)
     {
       p_n->value = p_n->value * 10 + d;
